+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 * Simplify the OpenGL performance setting
 * Add horizontal scrollbar in the input editor (#316)
 * Change the format of the input file, to make it more robust and adaptative (#332)
+* Input Editor window can get behind main window (#334, #49)
 
 ### Fixed